Added seeking, fleeing and flocking behavior to scripts
Scene now creates birds and lizards that seek each other and flock to a common point. This version also includes another scene that's unfinished due to some technical issues I ran into with the camera not rendering anything in the Scene or Game view. This put me behind on the independent seek/flee project and I currently have jellyfish in a scene that are set to seek and flee but are not quite finished. I have an image in Scenes of the current progress but it needs some work.